refactor: change variable type from long to int

+++ src/main/java/lotto/validator/PurchaseAmountValidator.java
@@ -4,14 +4,14 @@ import lotto.common.ErrorMessage;
 
 public class PurchaseAmountValidator {
 
-    public Long validatePurchaseAmount(String purchaseAmountInput) {
+    public int validatePurchaseAmount(String purchaseAmountInput) {
 
         if (purchaseAmountInput == null || purchaseAmountInput.isBlank()) {
             throw new IllegalArgumentException(ErrorMessage.NULL_OR_EMPTY.getMessage());
         }
 
         try {
-            long purchaseAmount = Long.parseLong(purchaseAmountInput);
+            int purchaseAmount = Integer.parseInt(purchaseAmountInput);
             if (purchaseAmount <= 0) {
                 throw new IllegalArgumentException(ErrorMessage.NOT_POSITIVE.getMessage());
             }
